Update styles and image upload functionality

+++ src/components/Blogs.jsx
@@ -1,45 +1,52 @@
 import React, { useEffect, useState } from "react";
 import BlogController from "./BlogController";
-import { auth,storage} from '../firebaseConfig';
+import { auth, storage } from '../firebaseConfig';
 import blogPageImg from "../assets/img/blogPageImg.png";
-import {ref, deleteObject } from "firebase/storage";
+import { ref, deleteObject } from "firebase/storage";
 
-const Blogs = ({ isLogin }) => {
+
+const Blogs = ({ isLogin,toast }) => {
     const [blogs, setBlogs] = useState([]);
 
     useEffect(() => {
         fetchAllBlogs();
     }, []);
 
-    const handleDelete = async (blogId, blogImgUrl,imgRefToFirebase) => { // Receive blogId as a parameter
-        try {
+    const handleDelete = async (blogId, blogImgUrl, imgRefToFirebase) => { // Receive blogId as a parameter
+        if (window.confirm("Are you sure you want to delete this blog?")) {
+            try {
+                toast.info("Deleting blog..");
+                const url = `https://personalwebsitebackend.onrender.com/blogs/remove/${blogId}`; // Use blogId in the URL
+                const response = await fetch(url, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json'
+                    }
+                });
 
-            const url = `https://personalwebsitebackend.onrender.com/blogs/remove/${blogId}`; // Use blogId in the URL
-            const response = await fetch(url, {
-                method: 'DELETE',
-                headers: {
-                    'Content-Type': 'application/json'
+                if (response.ok) {
+                    toast.success("Blog deleted successfully");
+                    console.log("Blog deleted, cleaning up resources..");
+                    if (blogImgUrl) {
+
+                        const storageRef = ref(storage);
+                        deleteObject(ref(storageRef, imgRefToFirebase)).then(() => {
+                            console.log("Previous image deleted");
+                        }).catch((error) => {
+                            toast.error("Blog deletion failed");
+                            console.error("Error deleting previous image:", error);
+                        });
+                    }
+
+                    fetchAllBlogs();
+                } else {
+                    toast.error("Blog deletion failed");
+                    throw new Error("Failed to delete blog");
                 }
-            });
-
-            if (response.ok) {
-                console.log("Blog deleted, cleaning up resources..");
-                if (blogImgUrl) {
-                   
-                    const storageRef = ref(storage);
-                    deleteObject(ref(storageRef,imgRefToFirebase )).then(() => {
-                        console.log("Previous image deleted");
-                    }).catch((error) => {
-                        console.error("Error deleting previous image:", error);
-                    });
-                }
-
-                fetchAllBlogs();
-            } else {
-                throw new Error("Failed to delete blog");
+            } catch (error) {
+                toast.error("Blog deletion failed");
+                console.error("Error deleting blog:", error);
             }
-        } catch (error) {
-            console.error("Error deleting blog:", error);
         }
     }
 
@@ -70,6 +77,7 @@ const Blogs = ({ isLogin }) => {
 
     return (
         <>
+            
             <div className="title-content max-width-1">
                 <div className="title-content-left">
                     <h1 className="blog-title-welcome">Welcome to Blogs!!</h1>
@@ -88,7 +96,7 @@ const Blogs = ({ isLogin }) => {
                         <div className="article-header">
                             <span className="article-title">{blog.title}</span>
                             {auth.currentUser && isLogin && auth.currentUser.displayName === blog.author && (
-                                <span><button onClick={() => handleDelete(blog._id, blog.imgUrl,blog.imgRefToFirebase)}>🗑️</button></span> // Pass blogId to handleDelete
+                                <span><button onClick={() => handleDelete(blog._id, blog.imgUrl, blog.imgRefToFirebase)}>🗑️</button></span> // Pass blogId to handleDelete
                             )}
                             <span className="article-right">
                                 <div className="article-author">@{blog.author}</div>
@@ -97,8 +105,11 @@ const Blogs = ({ isLogin }) => {
                         </div>
                         <div className="hr-nav m-auto"><hr /></div>
                         <div className="article-content">
-                            {blog.imgUrl && <img className="blogImage" src={blog.imgUrl} alt="blog" />}
-                            <p>{blog.content}</p>
+
+                            <p>
+                                {blog.imgUrl && <img className="blogImage" src={blog.imgUrl} alt="blog" />}
+                                {blog.content}
+                            </p>
                         </div>
                     </article>
                 ))}
